requirements.txt
gradio==4.36.1
groq==0.9.0
pytesseract==0.3.10
pymupdf==1.24.5
ej2-streamlit-grids==25.2.3
pillow=10.3.0
boto3==1.34.124
sentence-transformers==3.0.1
chromadb==0.5.0
langchain-community==0.2.4
uuid==1.30
langchain==0.2.3
Syncfusion.PDF.OCR.NET